pyproject.toml
[project]
name = "lang-chain-tutorial"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kang Seonghyeon", email = "shyeon.kang@gmail.com"},
]
dependencies = [
    "langsmith>=0.1.120",
    "openai>=1.45.0",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.2.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
